test: extract connection service white-box suite
